Remove `NavComponent` since `Navbar.js` no longer does routing
