+++ front-end/src/navbar/NavBar.js
@@ -1,20 +1,40 @@
-import NavComponent from './NavComponent.js';
-import Home from '../home/Home';
-import LookUp from '../lookup/LookUp.js';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import CreateEvents from '../create-events/CreateEvents.js';
+import { Link } from 'react-router-dom';
+import * as React from 'react';
+import Box from '@mui/material/Box';
+import CssBaseline from '@mui/material/CssBaseline';
+import BottomNavigation from '@mui/material/BottomNavigation';
+import BottomNavigationAction from '@mui/material/BottomNavigationAction';
+import Paper from '@mui/material/Paper';
+import AccountCircleIcon from '@mui/icons-material/AccountCircle';
+import SearchIcon from '@mui/icons-material/Search';
+import GroupsIcon from '@mui/icons-material/Groups';
+import PeopleIcon from '@mui/icons-material/People';
+import EventIcon from '@mui/icons-material/Event';
 
-const NavBar = props => {
+const NavComponent = props => {
+    const [value, setValue] = React.useState(0);
+    const ref = React.useRef(null);
     return (
-        <Router>
-            <NavComponent />
-            <Routes>
-                <Route path='/home' element={<Home />} />
-                <Route path='/lookup' element={<LookUp />} />
-                <Route path='/create-event' element={<CreateEvents />} />
-            </Routes>
-        </Router>
+        <Box sx={{ pb: 7 }} ref={ref}>
+            <CssBaseline />
+            <Paper sx={{ position: 'fixed', bottom: 0, left: 0, right: 0 }} elevation={3}>
+                <BottomNavigation
+                    showLabels
+                    value={value}
+                    onChange={(event, newValue) => {
+                        setValue(newValue);
+                    }}
+                >
+                    <BottomNavigationAction label="Profile" icon={<AccountCircleIcon />} component={Link} to="/home" />
+                    <BottomNavigationAction label="Look Up" icon={<SearchIcon />} component={Link} to="/lookup"/>
+                    <BottomNavigationAction label="Friends" icon={<PeopleIcon />} component={Link} to="/create-event"/>
+                    <BottomNavigationAction label="Groups" icon={<GroupsIcon />} />
+                    <BottomNavigationAction label="Events" icon={<EventIcon />} />
+
+                </BottomNavigation>
+            </Paper>
+        </Box>
     );
 }
 
-export default NavBar;
+export default NavComponent
